[backend] Rename DbUpdater package
