+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/listener/commits/DbUpdateException.java
@@ -1,4 +1,4 @@
-package de.aaaaaaah.velcom.backend.listener.dbupdate;
+package de.aaaaaaah.velcom.backend.listener.commits;
 
 /**
  * An exception thrown when the {@link DbUpdater} failed to update the database for a repo.
